* changes:
  fix(drivers/pl011): fix MISRA C:2012 Rule 10.4 in pl011.c
  fix(lib/realm): fix MISRA C:2012 Rule 18.2 in granule.c
  fix(rmm): fix MISRA C:2012 Rule 17.7
  fix(runtime): fix MISRA C:2012 Rule 10.5 in realm.c
  fix(lib/realm): fix MISRA C:2012 issues in buffer.c
  fix(lib/realm): fix MISRA C:2012 Rule 10.3 in status.h
  fix(lib/realm): fix MISRA C:2012 Rule 10.3 in granule.h
  fix(lib/measurement): fix MISRA C:2012 Rules 10.3 in measurement.c
  fix(lib/arch): fix MISRA C:2012 Rule 10.3 in EXTRACT usage
  fix(lib/arch): fix MISRA C:2012 Rule 10.3 in simd.c
  fix(drivers/pl011): fix MISRA C:2012 Rule 10.3 in pl011.c

+++ drivers/pl011/src/pl011.c
@@ -24,7 +24,6 @@ int uart_init(uintptr_t base_addr,
 	unsigned int div;
 
 	/* Check Base address, baud rate and UART clock for sanity */
-
 	if (base_addr == 0UL) {
 		return -1;
 	}
@@ -40,8 +39,8 @@ int uart_init(uintptr_t base_addr,
 	/* Disable UART before programming */
 	write32(0U, (void *)((RMM_UART_ADDR) + UARTCR));
 
-	/* Program the baudrate */
-	div = (uart_clk * 4)/baud_rate;
+	/* Program the baud rate */
+	div = (uart_clk * 4U) / baud_rate;
 
 	/* IBRD = Divisor >> 6 */
 	write32(div >> 6, (void *)((RMM_UART_ADDR) + UARTIBRD));
@@ -64,7 +63,7 @@ int uart_init(uintptr_t base_addr,
 static void uart_putc(char ch)
 {
 	uart_wait();
-	write8(ch, (void *)((RMM_UART_ADDR) + UARTDR));
+	write8((uint8_t)ch, (void *)((RMM_UART_ADDR) + UARTDR));
 }
 
 /* Serial output - called from printf */
